pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "boto3>=1.41.5",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-aws>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
]
